factor out the device dispatching code and implement CANTaskimplement CANTask and ActisenseTask
The project was originally to only do listening, and to avoid
dynamic dispatches to ports -- since messages on the NMEA2000 bus
are really low frequency. However, it turns out that the standard
NMEA2000 mechanisms to easily recognize to devices (instance numbers)
aren't supported by all devices, among which the fuel gauges we
have.

Since I don't want to shell out 100 eur for a programming device
each time someone feels like strongarming me, I've decided to implement
device enumeration on the bus, and allow to discriminate the devices
on their serial number. That's DeviceDispatcher